workspace/pyproject.toml
# ═══════════════════════════════════════════════════════════════════════════════
#                    MachineNativeOps Python Project Configuration
#                    Python 專案配置
# ═══════════════════════════════════════════════════════════════════════════════

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "machinenativeops"
version = "2.0.0"
description = "Machine-Native Operations Platform - 機器原生運維平台"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<3.13"
authors = [
    {name = "MachineNativeOps Team", email = "team@machinenativeops.io"}
]
maintainers = [
    {name = "MachineNativeOps Team", email = "team@machinenativeops.io"}
]
keywords = [
    "ai-agents",
    "multi-agent",
    "autonomous-systems",
    "automation",
    "orchestration",
    "langchain",
    "crewai",
    "autogen",
    "mcp",
    "devsecops",
    "slsa",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "jsonschema>=4.20.0",
    "typing-extensions>=4.8.0",
    # AI/ML Frameworks
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.10",
    "langgraph>=0.0.20",
    "openai>=1.6.0",
    # Multi-Agent
    "crewai>=0.1.0",
    "pyautogen>=0.2.0",
    # Async & Concurrency
    "asyncio>=3.4.3",
    "aiohttp>=3.9.0",
    "httpx>=0.26.0",
    # Monitoring & Observability
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    # Data Processing
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    # Security
    "cryptography>=41.0.0",
    "pyjwt>=2.8.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    "fastapi>=0.127.0",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.38.0",
    "setuptools>=80.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "ruff>=0.1.8",
    "black>=23.12.0",
    "isort>=5.13.0",
    "pre-commit>=3.6.0",
]
ml = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
]
full = [
    "machinenativeops[dev,ml]",
]

[project.urls]
Homepage = "https://github.com/MachineNativeOps/MachineNativeOps"
Documentation = "https://github.com/MachineNativeOps/MachineNativeOps/tree/main/docs"
Repository = "https://github.com/MachineNativeOps/MachineNativeOps.git"
Issues = "https://github.com/MachineNativeOps/MachineNativeOps/issues"
Changelog = "https://github.com/MachineNativeOps/MachineNativeOps/blob/main/CHANGELOG.md"

[project.scripts]
machinenativeops = "core.cli:main"
mnops = "core.cli:main"
machinenativeops-workflow = "automation.pipelines.instant_execution_pipeline:main"
machinenativeops-contract = "core.contract_engine:main"

[project.entry-points."machinenativeops.plugins"]
default = "core.plugins.default:DefaultPlugin"

# ═══════════════════════════════════════════════════════════════════════════════
#                         Tool Configurations
# ═══════════════════════════════════════════════════════════════════════════════

[tool.setuptools]
package-dir = {"" = "src"}
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["core", "core.*", "automation", "automation.*", "ai", "ai.*", "governance", "governance.*", "enterprise", "enterprise.*"]

[tool.setuptools.package-data]
"*" = ["py.typed", "*.yaml", "*.yml", "*.json", "*.txt"]

# Black Configuration
[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | node_modules
)/
'''

# isort Configuration
[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
skip = [".venv", "node_modules", "dist", "build"]
known_first_party = ["machinenativeops", "core", "automation", "ai", "governance", "enterprise"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

# Ruff Configuration
[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "W191",   # indentation contains tabs
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["ARG001"]

# MyPy Configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "crewai.*",
    "autogen.*",
    "torch.*",
    "transformers.*",
]
ignore_missing_imports = true

# Pytest Configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "-p no:warnings",
]
testpaths = [
    "src/tests",
]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage Configuration
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/conftest.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
fail_under = 80
show_missing = true
